backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Sheets Integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
playwright==1.40.0
httpx==0.25.1

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2 